requirements.txt
uvicorn>=0.16.0,<0.18.0
fastapi>=0.64.0,<1.0.0
requests>=2.24.0,<2.32.0
pydantic
pydantic-settings
numpy>=1.26.0,<2.0.0

# If locally copied whl file inside titanic_model_api then use given below
bike_sharing_model-0.1.0-py3-none-any.whl

# If uploded your package in PyPI then use the given below
#titanic-model==0.0.1